Get the id of the last model ran

This function returns the model ids for the last number models saved to leaderboard. This function is intended to be used with peak.

at_last(number = 1)

Arguments

number
number that indicates the last n models to peak at.

Value

model id of last models

Examples

# peak at previous model
board() %>%
  peak(at_last())

#> # A tibble: 10 x 13
#>     rank    id dir    model metric  score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    5.    13 model… rf    Accura… 0.957     NA boot     25.   12. <lis… <lis…
#>  2    6.    10 model… rf    Accura… 0.956     NA boot     25.    9. <lis… <lis…
#>  3    7.     1 model… rf    Accura… 0.955     NA boot     25.    1. <lis… <lis…
#>  4    8.    24 model… rf    Accura… 0.955     NA boot     25.   23. <lis… <lis…
#>  5    9.     4 model… rf    Accura… 0.955     NA boot     25.    4. <lis… <lis…
#>  6   10.    25 model… rf    Accura… 0.954     NA boot     25.   24. <lis… <lis…
#>  7   11.    15 model… rf    Accura… 0.954     NA boot     25.   14. <lis… <lis…
#>  8   12.    17 model… rf    Accura… 0.954     NA boot     25.   16. <lis… <lis…
#>  9   13.    12 model… rf    Accura… 0.953     NA boot     25.   11. <lis… <lis…
#> 10   14.    19 model… rf    Accura… 0.952     NA boot     25.   18. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

# peak at last two models
board() %>%
  peak(at_last(2))

#> # A tibble: 10 x 13
#>     rank    id dir    model metric  score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    4.     2 model… rf    Accura… 0.957     NA boot     25.    2. <lis… <lis…
#>  2    5.    13 model… rf    Accura… 0.957     NA boot     25.   12. <lis… <lis…
#>  3    6.    10 model… rf    Accura… 0.956     NA boot     25.    9. <lis… <lis…
#>  4    7.     1 model… rf    Accura… 0.955     NA boot     25.    1. <lis… <lis…
#>  5    8.    24 model… rf    Accura… 0.955     NA boot     25.   23. <lis… <lis…
#>  6    9.     4 model… rf    Accura… 0.955     NA boot     25.    4. <lis… <lis…
#>  7   10.    25 model… rf    Accura… 0.954     NA boot     25.   24. <lis… <lis…
#>  8   11.    15 model… rf    Accura… 0.954     NA boot     25.   14. <lis… <lis…
#>  9   12.    17 model… rf    Accura… 0.954     NA boot     25.   16. <lis… <lis…
#> 10   13.    12 model… rf    Accura… 0.953     NA boot     25.   11. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

# peak at last three models
# and model 1
board() %>%
  peak(at_last(3), 1)

#> # A tibble: 18 x 13
#>     rank    id dir    model metric  score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    7.     1 model… rf    Accura… 0.955     NA boot     25.    1. <lis… <lis…
#>  2    8.    24 model… rf    Accura… 0.955     NA boot     25.   23. <lis… <lis…
#>  3    9.     4 model… rf    Accura… 0.955     NA boot     25.    4. <lis… <lis…
#>  4   10.    25 model… rf    Accura… 0.954     NA boot     25.   24. <lis… <lis…
#>  5   11.    15 model… rf    Accura… 0.954     NA boot     25.   14. <lis… <lis…
#>  6   12.    17 model… rf    Accura… 0.954     NA boot     25.   16. <lis… <lis…
#>  7   13.    12 model… rf    Accura… 0.953     NA boot     25.   11. <lis… <lis…
#>  8   14.    19 model… rf    Accura… 0.952     NA boot     25.   18. <lis… <lis…
#>  9   15.    18 model… rf    Accura… 0.951     NA boot     25.   17. <lis… <lis…
#> 10   16.    16 model… rf    Accura… 0.951     NA boot     25.   15. <lis… <lis…
#> 11   17.    11 model… rf    Accura… 0.950     NA boot     25.   10. <lis… <lis…
#> 12   18.     9 model… rf    Accura… 0.949     NA boot     25.    8. <lis… <lis…
#> 13   19.     3 model… rf    Accura… 0.949     NA boot     25.    3. <lis… <lis…
#> 14   20.    20 model… rf    Accura… 0.948     NA boot     25.   19. <lis… <lis…
#> 15   21.     7 model… rf    Accura… 0.947     NA boot     25.    6. <lis… <lis…
#> 16   22.     6 model… rf    Accura… 0.946     NA boot     25.    5. <lis… <lis…
#> 17   23.     8 model… rf    Accura… 0.946     NA boot     25.    7. <lis… <lis…
#> 18   24.    23 model… rf    Accura… 0.944     NA boot     25.   22. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

board() %>%
  dplyr::filter(id %in% at_last(3))

#> # A tibble: 3 x 13
#>    rank    id dir    model metric  score public method   num group index  tune 
#>   <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <list> <lis>
#> 1    8.    24 model… rf    Accura… 0.955     NA boot     25.   23. <list… <lis…
#> 2   10.    25 model… rf    Accura… 0.954     NA boot     25.   24. <list… <lis…
#> 3   24.    23 model… rf    Accura… 0.944     NA boot     25.   22. <list… <lis…
#> # ... with 1 more variable: seeds <list>