A tibble leaderboard for caret train objects

This function updates and returns the model leaderboard for the project.

board(model = NULL, path = NULL, dir = NULL, save = TRUE,
  quiet = FALSE)

Arguments

model
model to add to the leaderboard. If no model is supplied (the default null), board returns the leaderboard tibble for the project.

path
globally sets the path to save models and leaderboards. By default, the path is the project directory found by. For best results, the path string should be constructed with file.path or here::here().

dir
globally sets name of directory where models are saved. This will be a subdirectory of the specified path. The default directory is named models_one. If no argument is supplied, the model will be saved in the previously specified directory. See the example below.

save
whether board should save the supplied model to dir. If FALSE the model will not be saved, but will be added to the leaderboard.

quiet
whether board should return the leaderboard tibble to the console. By default quiet = FALSE means that the tibble prints to console. quiet = TRUE is useful in a .Rmd environment, where you want to add the model to the leaderboard without printing the tibble.

Value

tibble containing the most up-to-date leaderboard.

Examples

# add caret model to leaderboard
# model saved to "models_one"
model <- train(...)

#> Error in eval(expr, envir, enclos): '...' used in an incorrect context

leadr::board(model)

#> # A tibble: 6 x 13
#>    rank    id dir    model  metric  score public method   num group index tune 
#>   <dbl>  <id> <chr>  <chr>  <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#> 1    1.     2 model… glmnet Accura… 0.973     NA cv       10.    1. <lis… <lis…
#> 2    1.     4 model… knn    Accura… 0.973     NA cv       10.    1. <lis… <lis…
#> 3    1.     6 model… glmnet Accura… 0.973     NA cv        5.    2. <lis… <lis…
#> 4    4.     5 ensem… rf     Accura… 0.960     NA cv       10.    1. <lis… <lis…
#> 5    5.     1 model… rf     Accura… 0.953     NA cv       10.    1. <lis… <lis…
#> 6    6.     3 model… rpart  Accura… 0.947     NA cv       10.    1. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

# return tibble leaderboard
leadr::board()

#> # A tibble: 6 x 13
#>    rank    id dir    model  metric  score public method   num group index tune 
#>   <dbl>  <id> <chr>  <chr>  <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#> 1    1.     2 model… glmnet Accura… 0.973     NA cv       10.    1. <lis… <lis…
#> 2    1.     4 model… knn    Accura… 0.973     NA cv       10.    1. <lis… <lis…
#> 3    1.     6 model… glmnet Accura… 0.973     NA cv        5.    2. <lis… <lis…
#> 4    4.     5 ensem… rf     Accura… 0.960     NA cv       10.    1. <lis… <lis…
#> 5    5.     1 model… rf     Accura… 0.953     NA cv       10.    1. <lis… <lis…
#> 6    6.     3 model… rpart  Accura… 0.947     NA cv       10.    1. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

# save to different directory
ensemble <- train(...)

#> Error in eval(expr, envir, enclos): '...' used in an incorrect context

leadr::board(ensemble, dir = "ensembles_one")

#> Error in leadr::board(ensemble, dir = "ensembles_one"): object 'ensemble' not found

# board automatically saves to previous directory
# model saved to "ensembles_one"
ensemble2 <- train(...)

#> Error in eval(expr, envir, enclos): '...' used in an incorrect context

leadr::board(ensembles2)

#> Error in leadr::board(ensembles2): object 'ensembles2' not found