Set seed for caret's trainControl

Given a starting seed and number of folds (or resamples), this function creates a vector of seeds needed for the `caret` trainControl object.

caret_seed(seed = 1, number)

Arguments

seed
starting seed for list of seeds

number
number of folds or resamples

Value

List of seeds used by trainControl.

Examples

trainControl(method = 'cv', number = 10, seeds = caret_seed(1234, 10))

#> $method
#> [1] "cv"
#> 
#> $number
#> [1] 10
#> 
#> $repeats
#> [1] NA
#> 
#> $search
#> [1] "grid"
#> 
#> $p
#> [1] 0.75
#> 
#> $initialWindow
#> NULL
#> 
#> $horizon
#> [1] 1
#> 
#> $fixedWindow
#> [1] TRUE
#> 
#> $skip
#> [1] 0
#> 
#> $verboseIter
#> [1] FALSE
#> 
#> $returnData
#> [1] TRUE
#> 
#> $returnResamp
#> [1] "final"
#> 
#> $savePredictions
#> [1] FALSE
#> 
#> $classProbs
#> [1] FALSE
#> 
#> $summaryFunction
#> function (data, lev = NULL, model = NULL) 
#> {
#>     if (is.character(data$obs)) 
#>         data$obs <- factor(data$obs, levels = lev)
#>     postResample(data[, "pred"], data[, "obs"])
#> }
#> <environment: namespace:caret>
#> 
#> $selectionFunction
#> [1] "best"
#> 
#> $preProcOptions
#> $preProcOptions$thresh
#> [1] 0.95
#> 
#> $preProcOptions$ICAcomp
#> [1] 3
#> 
#> $preProcOptions$k
#> [1] 5
#> 
#> $preProcOptions$freqCut
#> [1] 19
#> 
#> $preProcOptions$uniqueCut
#> [1] 10
#> 
#> $preProcOptions$cutoff
#> [1] 0.9
#> 
#> 
#> $sampling
#> NULL
#> 
#> $index
#> NULL
#> 
#> $indexOut
#> NULL
#> 
#> $indexFinal
#> NULL
#> 
#> $timingSamps
#> [1] 0
#> 
#> $predictionBounds
#> [1] FALSE FALSE
#> 
#> $seeds
#> $seeds[[1]]
#> [1] 114 622 609 999
#> 
#> $seeds[[2]]
#> [1] 861 640  10 232
#> 
#> $seeds[[3]]
#> [1] 667 514 693 544
#> 
#> $seeds[[4]]
#> [1] 283 923 292 835
#> 
#> $seeds[[5]]
#> [1] 287 267 187 232
#> 
#> $seeds[[6]]
#> [1] 317 303 159  40
#> 
#> $seeds[[7]]
#> [1] 219 810 525 912
#> 
#> $seeds[[8]]
#> [1] 832  46 456 265
#> 
#> $seeds[[9]]
#> [1] 305 507 181 758
#> 
#> $seeds[[10]]
#> [1] 202 259 991 805
#> 
#> $seeds[[11]]
#> [1] 554
#> 
#> 
#> $adaptive
#> $adaptive$min
#> [1] 5
#> 
#> $adaptive$alpha
#> [1] 0.05
#> 
#> $adaptive$method
#> [1] "gls"
#> 
#> $adaptive$complete
#> [1] TRUE
#> 
#> 
#> $trim
#> [1] FALSE
#> 
#> $allowParallel
#> [1] TRUE
#>