Peak at leaderboard positions

This function returns a tibble with showing the models ranked closest to the supplied model. Usually chained with the magrittr pipe %>%.

peak(leadrboard, ..., n = 10, how = "centered")

Arguments

leadrboard
leaderboard tibble returned by board

...
a number, numbers, or vector of numbers to that correspond to model ids in the leaderboard. See at_last to easily specify model ids.

n
the number of rows to return. By default, the tibble will return 10 rows. peak will return a tibble with more than 20 rows, but the console output is limited by options(tibble.print_max), which has a default of 20.

how
takes values of c("centered", "above", "below"). Determines where to peak around the first number supplied to ....

Value

A subset of the leaderboard tibble. Given the supplied model ids, peak tries to return a tibble of length n. If the maximum distance between model ids are greater than n, peak will return the shortest tibble that includes all the models. In supporting consoles, the supplied model ids are highlighted.

Examples

# peak at last model saved
board() %>%
  peak(at_last())

#> # A tibble: 10 x 13
#>     rank    id dir    model  metric score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr>  <chr>  <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    1.    26 model… glmnet Accur… 0.973     NA cv        5.   25. <lis… <lis…
#>  2    2.     5 model… glmnet Accur… 0.973     NA boot     25.    1. <lis… <lis…
#>  3    3.     1 model… rf     Accur… 0.960     NA boot     25.    1. <lis… <lis…
#>  4    4.     6 model… rf     Accur… 0.957     NA boot     25.    5. <lis… <lis…
#>  5    5.     3 model… rf     Accur… 0.957     NA boot     25.    3. <lis… <lis…
#>  6    6.    20 model… rf     Accur… 0.955     NA boot     25.   19. <lis… <lis…
#>  7    7.     4 model… rf     Accur… 0.955     NA boot     25.    4. <lis… <lis…
#>  8    8.    25 model… rf     Accur… 0.954     NA boot     25.   24. <lis… <lis…
#>  9    9.    15 model… rf     Accur… 0.953     NA boot     25.   14. <lis… <lis…
#> 10   10.     9 model… rf     Accur… 0.953     NA boot     25.    8. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

# peak at models 1 and 2
board() %>%
  peak(1, 2)

#> # A tibble: 17 x 13
#>     rank    id dir    model metric  score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    3.     1 model… rf    Accura… 0.960     NA boot     25.    1. <lis… <lis…
#>  2    4.     6 model… rf    Accura… 0.957     NA boot     25.    5. <lis… <lis…
#>  3    5.     3 model… rf    Accura… 0.957     NA boot     25.    3. <lis… <lis…
#>  4    6.    20 model… rf    Accura… 0.955     NA boot     25.   19. <lis… <lis…
#>  5    7.     4 model… rf    Accura… 0.955     NA boot     25.    4. <lis… <lis…
#>  6    8.    25 model… rf    Accura… 0.954     NA boot     25.   24. <lis… <lis…
#>  7    9.    15 model… rf    Accura… 0.953     NA boot     25.   14. <lis… <lis…
#>  8   10.     9 model… rf    Accura… 0.953     NA boot     25.    8. <lis… <lis…
#>  9   11.    18 model… rf    Accura… 0.953     NA boot     25.   17. <lis… <lis…
#> 10   12.    22 model… rf    Accura… 0.952     NA boot     25.   21. <lis… <lis…
#> 11   13.    17 model… rf    Accura… 0.952     NA boot     25.   16. <lis… <lis…
#> 12   14.    21 model… rf    Accura… 0.951     NA boot     25.   20. <lis… <lis…
#> 13   15.    23 model… rf    Accura… 0.950     NA boot     25.   22. <lis… <lis…
#> 14   16.    14 model… rf    Accura… 0.949     NA boot     25.   13. <lis… <lis…
#> 15   17.    16 model… rf    Accura… 0.949     NA boot     25.   15. <lis… <lis…
#> 16   18.    13 model… rf    Accura… 0.948     NA boot     25.   12. <lis… <lis…
#> 17   19.     2 model… rf    Accura… 0.948     NA boot     25.    2. <lis… <lis…
#> # ... with 1 more variable: seeds <list>

board() %>%
  peak(c(1, 2))

#> # A tibble: 17 x 13
#>     rank    id dir    model metric  score public method   num group index tune 
#>    <dbl>  <id> <chr>  <chr> <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <lis> <lis>
#>  1    3.     1 model… rf    Accura… 0.960     NA boot     25.    1. <lis… <lis…
#>  2    4.     6 model… rf    Accura… 0.957     NA boot     25.    5. <lis… <lis…
#>  3    5.     3 model… rf    Accura… 0.957     NA boot     25.    3. <lis… <lis…
#>  4    6.    20 model… rf    Accura… 0.955     NA boot     25.   19. <lis… <lis…
#>  5    7.     4 model… rf    Accura… 0.955     NA boot     25.    4. <lis… <lis…
#>  6    8.    25 model… rf    Accura… 0.954     NA boot     25.   24. <lis… <lis…
#>  7    9.    15 model… rf    Accura… 0.953     NA boot     25.   14. <lis… <lis…
#>  8   10.     9 model… rf    Accura… 0.953     NA boot     25.    8. <lis… <lis…
#>  9   11.    18 model… rf    Accura… 0.953     NA boot     25.   17. <lis… <lis…
#> 10   12.    22 model… rf    Accura… 0.952     NA boot     25.   21. <lis… <lis…
#> 11   13.    17 model… rf    Accura… 0.952     NA boot     25.   16. <lis… <lis…
#> 12   14.    21 model… rf    Accura… 0.951     NA boot     25.   20. <lis… <lis…
#> 13   15.    23 model… rf    Accura… 0.950     NA boot     25.   22. <lis… <lis…
#> 14   16.    14 model… rf    Accura… 0.949     NA boot     25.   13. <lis… <lis…
#> 15   17.    16 model… rf    Accura… 0.949     NA boot     25.   15. <lis… <lis…
#> 16   18.    13 model… rf    Accura… 0.948     NA boot     25.   12. <lis… <lis…
#> 17   19.     2 model… rf    Accura… 0.948     NA boot     25.    2. <lis… <lis…
#> # ... with 1 more variable: seeds <list>